requirements.txt
# Requires Python >=3.9 and <3.13
matplotlib==3.10.3
numpy==2.2.6
pandas==2.2.3
pillow==11.2.1
pygame==2.6.1
scipy==1.15.3
sounddevice==0.5.2
soundfile==0.13.1
